reverted to 87ee2ed73ca2d4ade44807082d44fd687455205b, working commit

+++ bottomLayer/src/routes/private/user.ts
@@ -1,23 +1,14 @@
 import express, { type Request, type Response } from 'express';
 import { pool } from '../../utils/sqlImport';
-import {
-  responseCallbackDelete,
-  responseCallbackPost,
-  responseCallbackUpdate,
-  responseCallbackGet
-} from '../../utils/responseCallback';
+import { responseCallbackDelete, responseCallbackPost, responseCallbackUpdate, responseCallbackGet } from '../../utils/responseCallback';
 import { checkAuthenticated } from '../../middleware/auth';
-import { convertImage } from '../../s3/convert-image';
 const router = express.Router();
 
 router.get('/', (req: Request, res: Response): void => {
   const userId = req.user;
   const getUser = async (): Promise<void> => {
     try {
-      const user = await pool.query(
-        'SELECT uid, first_name, last_name, email, username, profile_picture FROM backend_schema.user WHERE uid = $1',
-        [userId]
-      );
+      const user = await pool.query('SELECT uid, first_name, last_name, email, username, profile_picture FROM backend_schema.user WHERE uid = $1', [userId]);
       const result = user.rows[0];
 
       responseCallbackGet(null, result, res, 'User');
@@ -29,7 +20,7 @@ router.get('/', (req: Request, res: Response): void => {
   void getUser();
 });
 // Endpoint for creating a specific user
-router.post('/', (req: Request, res: Response) => {
+router.post('/', checkAuthenticated, (req: Request, res: Response) => {
   const {
     first_name,
     last_name,
@@ -42,17 +33,15 @@ router.post('/', (req: Request, res: Response) => {
     following
   } = req.body;
 
-  console.log(req.body);
-
   const insertUser = async (): Promise<void> => {
     try {
-      const URL = await convertImage(profile_picture, username, false);
       await pool.query(`
       INSERT INTO backend_schema.user (
         first_name, last_name, email, username, password, private_option, followers, following, profile_picture
         ) VALUES ( 
           $1, $2, $3, $4, $5, $6, $7, $8, $9)`,
-      [first_name, last_name, email, username, password, private_option, followers, following, URL]);
+      [first_name, last_name, email, username, password, private_option, followers, following, profile_picture]);
+
       responseCallbackPost(null, res, 'User');
     } catch (error) {
       responseCallbackPost(error, res);
@@ -69,10 +58,7 @@ router.delete('/', checkAuthenticated, (req: Request, res: Response): void => {
 
   const deleteUser = async (): Promise<void> => {
     try {
-      const deleteUser = await pool.query(
-        'DELETE FROM backend_schema.user WHERE uid = $1',
-        [userId]
-      );
+      const deleteUser = await pool.query('DELETE FROM backend_schema.user WHERE uid = $1', [userId]);
       responseCallbackDelete(null, userId, res, 'User', deleteUser.rowCount);
     } catch (error) {
       responseCallbackDelete(error, userId, res, 'User');
@@ -101,7 +87,6 @@ router.put('/', checkAuthenticated, (req: Request, res: Response): void => {
   } = req.body;
   const updateUser = async (): Promise<void> => {
     try {
-      const URL = await convertImage(profile_picture, username, false);
       const updateUser = await pool.query(`UPDATE backend_schema.user
         SET first_name = $1,
             last_name = $2,
@@ -113,19 +98,7 @@ router.put('/', checkAuthenticated, (req: Request, res: Response): void => {
             following = $8,
             profile_picture = $9
         WHERE uid = $10`,
-        [
-          first_name,
-          last_name,
-          email,
-          username,
-          password,
-          private_option,
-          followers,
-          following,
-          profile_picture,
-          userId
-        ]
-      );
+      [first_name, last_name, email, username, password, private_option, followers, following, profile_picture, userId]);
       // responds with successful update even when no changes are made
       responseCallbackUpdate(null, userId, res, 'User', updateUser.rowCount);
     } catch (error) {
